requirements.txt
streamlit>=1.44,<2.0
duckdb>=1.1,<2.0
pandas>=2.2,<3.0
pyarrow>=18.0,<20.0
requests>=2.32,<3.0
azure-storage-blob>=12.23,<13.0
pytest>=8.0,<9.0
